refactor: extract notifications hook and item
